requirements.txt
a2wsgi==1.10.10
aiosmtplib==5.1.0
aiosqlite==0.21.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
apache-airflow==3.2.0
apache-airflow-core==3.2.0
apache-airflow-providers-common-compat==1.14.2
apache-airflow-providers-common-io==1.7.2
apache-airflow-providers-common-sql==1.34.0
apache-airflow-providers-smtp==2.4.4
apache-airflow-providers-standard==1.12.2
apache-airflow-task-sdk==1.2.0
argcomplete==3.6.3
asgiref==3.11.1
attrs==26.1.0
babel==2.18.0
cadwyn==6.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
colorlog==6.10.1
cron-descriptor==2.0.8
croniter==6.2.2
cryptography==46.0.7
deprecated==1.3.1
dill==0.4.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.3
fastapi-cli==0.0.24
fsspec==2026.3.0
googleapis-common-protos==1.74.0
greenback==1.3.0
greenlet==3.4.0
grpcio==1.80.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib-metadata==8.7.1
itsdangerous==2.2.0
jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
libcst==1.8.6
linkify-it-py==2.1.0
lockfile==0.12.2
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
methodtools==0.4.7
more-itertools==11.0.2
msgspec==0.21.0
natsort==8.4.0
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-exporter-otlp-proto-http==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
outcome==1.3.0.post0
packaging==26.0
pathlib-abc==0.5.2
pathspec==1.0.4
pendulum==3.2.0
pluggy==1.6.0
protobuf==6.33.6
psutil==7.2.2
py4j==0.10.9.9
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-extra-types==2.11.2
pydantic-settings==2.13.1
pygments==2.20.0
pygtrie==2.5.0
pyjwt==2.12.1
pyspark==4.1.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
python-slugify==8.0.4
pyyaml==6.0.3
pyyaml-ft==8.0.0
referencing==0.37.0
requests==2.33.1
rich==14.3.4
rich-argparse==1.7.2
rich-toolkit==0.19.7
rpds-py==0.30.0
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.49
sqlparse==0.5.5
starlette==0.52.1
structlog==25.5.0
svcs==25.1.0
tabulate==0.10.0
tenacity==9.1.4
termcolor==3.3.0
text-unidecode==1.3
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2026.1
uc-micro-py==2.0.0
universal-pathlib==0.3.10
urllib3==2.6.3
uuid6==2025.0.1
uvicorn==0.44.0
watchfiles==1.1.1
websockets==16.0
wirerope==1.0.0
wrapt==2.1.2
zipp==3.23.0
